+++ test/thunkFactory.spec.js
@@ -1,4 +1,5 @@
 import {
+  deletePlan,
   getPlan,
   getPlans,
   postPlan,
@@ -159,6 +160,8 @@ describe('Thunk Factory', () => {
     const store = mockStore({
       plans: {
         list: [],
+        page: 1,
+        filter: {},
       },
     });
 
@@ -307,6 +310,86 @@ describe('Thunk Factory', () => {
       });
   });
 
+  it('should reload resources when clear resource sort succeed', () => {
+    const store = mockStore({
+      plans: {
+        list: [],
+        page: 1,
+      },
+    });
+
+    const mockData = {
+      data: {
+        data: [{ name: 'Finish off' }],
+        page: 1,
+        pages: 1,
+        total: 1,
+      },
+    };
+    getPlans.mockResolvedValueOnce(mockData);
+
+    const planThunks = createThunkFor('plans');
+    const expectedActions = [
+      { type: 'plan/clearPlansSort', payload: undefined },
+      { type: 'plan/getPlansRequest', payload: undefined },
+      { type: 'plan/getPlansSuccess', payload: mockData },
+    ];
+
+    const onSuccess = jest.fn();
+    const onError = jest.fn();
+
+    return store
+      .dispatch(planThunks.clearPlansSort(onSuccess, onError))
+      .then(() => {
+        expect(store.getActions()).toEqual(expectedActions);
+        expect(onSuccess).toHaveBeenCalledTimes(1);
+        expect(onError).toHaveBeenCalledTimes(0);
+        expect(getPlans).toHaveBeenCalledWith({ page: 1 });
+      });
+  });
+
+  it('should dispatch error action when clear resource sort fails', () => {
+    const store = mockStore({
+      plans: {
+        list: [],
+        error: null,
+        page: 1,
+      },
+    });
+
+    const error = {
+      status: 404,
+      code: 404,
+      name: 'Error',
+      message: 'Not Found',
+      developerMessage: 'Not Found',
+      userMessage: 'Not Found',
+      error: 'Error',
+      error_description: 'Not Found',
+    };
+
+    getPlans.mockRejectedValueOnce(error);
+
+    const planThunks = createThunkFor('plans');
+    const expectedActions = [
+      { type: 'plan/clearPlansSort', payload: undefined },
+      { type: 'plan/getPlansRequest', payload: undefined },
+      { type: 'plan/getPlansFailure', payload: error },
+    ];
+
+    const onSuccess = jest.fn();
+    const onError = jest.fn();
+
+    return store
+      .dispatch(planThunks.clearPlansSort(onSuccess, onError))
+      .then(() => {
+        expect(store.getActions()).toEqual(expectedActions);
+        expect(onSuccess).toHaveBeenCalledTimes(0);
+        expect(onError).toHaveBeenCalledTimes(1);
+        expect(getPlans).toHaveBeenCalledWith({ page: 1 });
+      });
+  });
+
   it('should dispatch required actions when search resources succeed', () => {
     const store = mockStore({
       plans: {
@@ -384,6 +467,7 @@ describe('Thunk Factory', () => {
     const store = mockStore({
       plans: {
         list: [],
+        page: 1,
       },
     });
 
@@ -413,6 +497,7 @@ describe('Thunk Factory', () => {
         expect(store.getActions()).toEqual(expectedActions);
         expect(onSuccess).toHaveBeenCalledTimes(1);
         expect(onError).toHaveBeenCalledTimes(0);
+        expect(getPlans).toHaveBeenCalledWith({ page: 1, sort: { name: -1 } });
       });
   });
 
@@ -420,6 +505,7 @@ describe('Thunk Factory', () => {
     const store = mockStore({
       plans: {
         list: [],
+        page: 1,
       },
     });
 
@@ -451,6 +537,7 @@ describe('Thunk Factory', () => {
         expect(store.getActions()).toEqual(expectedActions);
         expect(onSuccess).toHaveBeenCalledTimes(0);
         expect(onError).toHaveBeenCalledTimes(1);
+        expect(getPlans).toHaveBeenCalledWith({ page: 1, sort: { name: -1 } });
       });
   });
 
@@ -458,6 +545,7 @@ describe('Thunk Factory', () => {
     const store = mockStore({
       plans: {
         list: [],
+        filter: {},
       },
     });
 
@@ -486,6 +574,7 @@ describe('Thunk Factory', () => {
         expect(store.getActions()).toEqual(expectedActions);
         expect(onSuccess).toHaveBeenCalledTimes(1);
         expect(onError).toHaveBeenCalledTimes(0);
+        expect(getPlans).toHaveBeenCalledWith({ page: 1, filter: {} });
       });
   });
 
@@ -737,6 +826,86 @@ describe('Thunk Factory', () => {
 
     return store
       .dispatch(planThunks.putPlan({}, onSuccess, onError))
+      .then(() => {
+        expect(store.getActions()).toEqual(expectedActions);
+        expect(onSuccess).toHaveBeenCalledTimes(0);
+        expect(onError).toHaveBeenCalledTimes(1);
+      });
+  });
+
+  it('should dispatch required actions when delete resource succeed', () => {
+    const store = mockStore({
+      plans: {
+        list: [],
+      },
+    });
+
+    const mockData = {
+      name: 'Finish off',
+    };
+
+    const mockGetData = {
+      data: {
+        data: [{ name: 'Finish off' }],
+        page: 1,
+        pages: 1,
+        total: 1,
+      },
+    };
+
+    const onSuccess = jest.fn();
+    const onError = jest.fn();
+
+    deletePlan.mockResolvedValueOnce(mockData);
+    getPlans.mockResolvedValueOnce(mockGetData);
+
+    const planThunks = createThunkFor('plans');
+    const expectedActions = [
+      { type: 'plan/deletePlanRequest', payload: undefined },
+      { type: 'plan/deletePlanSuccess', payload: mockData },
+      { type: 'plan/getPlansRequest', payload: undefined },
+      { type: 'plan/getPlansSuccess', payload: mockGetData },
+    ];
+
+    return store
+      .dispatch(planThunks.deletePlan({}, onSuccess, onError))
+      .then(() => {
+        expect(store.getActions()).toEqual(expectedActions);
+        expect(onSuccess).toHaveBeenCalledTimes(1);
+        expect(onError).toHaveBeenCalledTimes(0);
+      });
+  });
+
+  it('should dispatch required actions when delete resource fails', () => {
+    const store = mockStore({
+      plans: {
+        list: [],
+      },
+    });
+
+    const error = {
+      status: 404,
+      code: 404,
+      name: 'Error',
+      message: 'Not Found',
+      developerMessage: 'Not Found',
+      userMessage: 'Not Found',
+      error: 'Error',
+      error_description: 'Not Found',
+    };
+
+    deletePlan.mockRejectedValueOnce(error);
+
+    const planThunks = createThunkFor('plans');
+    const expectedActions = [
+      { type: 'plan/deletePlanRequest', payload: undefined },
+      { type: 'plan/deletePlanFailure', payload: error },
+    ];
+    const onSuccess = jest.fn();
+    const onError = jest.fn();
+
+    return store
+      .dispatch(planThunks.deletePlan({}, onSuccess, onError))
       .then(() => {
         expect(store.getActions()).toEqual(expectedActions);
         expect(onSuccess).toHaveBeenCalledTimes(0);